(SLV-769) Move provision_pe_xl to helper
This commit moves the `provision_pe_xl_nodes.rb` utility to a
`pe_xl_helper.rb` file.  This file is then required in the Rakefile.
